requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
pycryptodome==3.18.0
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0